+++ nengo/builder/node.py
@@ -39,7 +39,7 @@ def build_node(model, node):
         sig_out = sig_in
     elif isinstance(node.output, Process):
         sig_out = Signal(shape=node.size_out, name="%s.out" % node)
-        model.build(node.output, sig_in, sig_out)
+        model.build(node.output, sig_in, sig_out, mode="set")
     elif callable(node.output):
         sig_out = (Signal(shape=node.size_out, name="%s.out" % node)
                    if node.size_out > 0 else None)